minor styling stuff
