+++ react/src/components/upload/UploadImage.js
@@ -1,6 +1,5 @@
 import axios from 'axios'
 
-// const maskAPI = 'https://hirsch-cs121-demo.herokuapp.com'
 const maskAPI = 'https://facee-309423.wl.r.appspot.com'
 
 export default function handleUpload(updateImage, updateMasks) {
